Change wind direction to int
As per OpenMeteo API response

+++ WeatherAnalyzer/Weather/Wind.cs
@@ -2,4 +2,4 @@ using WeatherAnalyzer.Util;
 
 namespace WeatherAnalyzer.Weather;
 
-public record struct Wind(ValueUnit<float> Speed, ValueUnit<float> Direction);
+public record struct Wind(ValueUnit<float> Speed, ValueUnit<int> Direction);
